apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
jsonschema>=4.23.0
openai>=1.58.0
azure-identity>=1.19.0
azure-keyvault-secrets>=4.9.0
azure-monitor-opentelemetry>=1.6.0
opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.1
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
black>=24.10.0
